(fix): 1- remove order page 2- make a customs checkout_page.dart 3- edit name checkout_view to checkout_view_body.dart
