+++ lib/presentation/checkout_page/page/checkout_page.dart
@@ -1,5 +1,4 @@
-import 'package:flowery_app/domain/entities/cart/cart_item_entity/cart_item_entity.dart';
-import 'package:flowery_app/presentation/checkout_page/view/checkout_view.dart';
+import 'package:flowery_app/presentation/checkout_page/view/checkout_view_body.dart';
 import 'package:flowery_app/presentation/checkout_page/view_model/adress_cubit/adress_cubit.dart';
 import 'package:flowery_app/presentation/checkout_page/view_model/checkout_cubit/checkout_view_model.dart';
 import 'package:flowery_app/presentation/checkout_page/view_model/payment_cubit/payment_cubit.dart';
@@ -8,8 +7,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 
 class CheckoutPage extends StatelessWidget {
   final int subTotal;
-  final List<CartItemEntity> cartItem;
-  const CheckoutPage({super.key, required this.subTotal, required this.cartItem});
+  const CheckoutPage({super.key, required this.subTotal});
 
   @override
   Widget build(BuildContext context) {
@@ -19,9 +17,8 @@ class CheckoutPage extends StatelessWidget {
         BlocProvider(create: (_) => PaymentCubit()),
         BlocProvider(create: (_) => AddressCubit()),
       ],
-      child: CheckoutView(
+      child: CheckoutViewBody(
         subTotal: subTotal,
-        cartItems: cartItem,
       ),
     );
   }
