+++ lib/utils/common_widgets/checkout_common_widgets/custom_divider.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+class CustomDivider extends StatelessWidget {
+  const CustomDivider({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: double.infinity,
+      height: 24,
+      child: Container(
+        decoration: const BoxDecoration(
+            color: Color(0xffEAEAEA)
+        ),
+      ),
+    );
+  }
+}
